Best Lift Management System
There is a building with 1000 floors and 2 lifts.
Which lift system is has the best performance and least usage over the long run.

Case 1:
The building is divided in 2 halves.
Floors 1 - 500 and 500 - 1000
Lift A is used for the lower half, and Lift B for the upper Half.

Case 2:
Lift A is at Floor 1 and Lift B is at Floor 1000.
The lift closest to the user reaches the user. If both lifts are equidistant, Lift which is lower moves

Case 3:
IT is the same case as Case 1 except, the initial positions of the Lifts are at 250 and 750 respectively.